feat:elapsed time on main

+++ mpi-implementation/include/tree_reduce.h
@@ -22,7 +22,7 @@
  *  @param up upper bound of data.
  *
  */
-void initialize_data_array(int rank, int *data, int n, size_t lb, size_t ub);
+void initialize_data_array(int rank, int *data, size_t n, size_t lb, size_t ub);
 
 /**
  *  @brief Scatter variable-sized segments across ranks using MPI_Scatterv.
@@ -58,7 +58,7 @@ int *distribute_data_array(
     int *displs,
     int local_n,
     int rank, 
-    int buf_size,
+    size_t buf_size,
     MPI_Comm comm
 );
