+++ mpi-implementation/src/tree_reduce.c
@@ -12,7 +12,7 @@
 void initialize_data_array(
     int rank,
     int *data,
-    int n,
+    size_t n,
     size_t lb,
     size_t ub)
 {
@@ -51,7 +51,7 @@ int *distribute_data_array(
     int *displs,
     int local_n,
     int rank, 
-    int buf_size,
+    size_t buf_size,
     MPI_Comm comm
 )
 {
